nicintel_spi: fix truncated MMIO BAR on 64-bit systems
nicintel_spi stored BAR0 into a 32-bit variable. On systems where the NIC
gets a 64-bit MMIO BAR above 4G, this truncates the address and causes
/dev/mem mmap failures (EPERM) and crashes.

Use uintptr_t for the MMIO base address so 64-bit BARs are handled correctly.
Tested by reading M25P40 (512 kB) on Intel 82599 (8086:10fb).

Reviewed-on: https://review.coreboot.org/c/flashrom/+/90638
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ nicintel_spi.c
@@ -285,7 +285,7 @@ static int nicintel_spi_init(const struct programmer_cfg *cfg)
 	if (!dev)
 		return 1;
 
-	uint32_t io_base_addr = pcidev_readbar(dev, PCI_BASE_ADDRESS_0);
+	uintptr_t io_base_addr = pcidev_readbar(dev, PCI_BASE_ADDRESS_0);
 	if (!io_base_addr)
 		return 1;
